Improved watcher

+++ start.sh
@@ -4,8 +4,15 @@ go run main.go &
 
 inotifywait -q -m -e close_write *.go **/*.go |
 while read -r file event; do
-    kill -9 $!
-    kill $(ps | grep main | awk '{print $1}')
+    if [ "$!" ] 
+    then
+        kill $!
+    fi
+
+    if [ "$(ps | grep main | awk '{print $1}')" ]
+    then 
+        kill $(ps | grep main | awk '{print $1}')
+    fi
     echo "Recompiling..."
     go run main.go &
 done
